+++ lib/AppInfo/Application.php
@@ -25,13 +25,16 @@
 namespace OCA\Polls\AppInfo;
 
 use Closure;
+// use OC\EventDispatcher\SymfonyAdapter;
 use OCP\AppFramework\App;
 use OCP\AppFramework\Bootstrap\IBootContext;
 use OCP\AppFramework\Bootstrap\IBootstrap;
 use OCP\AppFramework\Bootstrap\IRegistrationContext;
+use OCP\Collaboration\Resources\IProviderManager;
 use OCP\Notification\IManager as NotificationManager;
 use OCP\Group\Events\GroupDeletedEvent;
 use OCP\User\Events\UserDeletedEvent;
+use OCP\Util;
 use OCA\Polls\Event\CommentAddEvent;
 use OCA\Polls\Event\CommentDeleteEvent;
 use OCA\Polls\Event\OptionConfirmedEvent;
@@ -62,6 +65,8 @@ use OCA\Polls\Listener\OptionListener;
 use OCA\Polls\Listener\PollListener;
 use OCA\Polls\Listener\ShareListener;
 use OCA\Polls\Listener\VoteListener;
+use OCA\Polls\Provider\ResourceProvider;
+use OCA\Polls\Provider\SearchProvider;
 
 class Application extends App implements IBootstrap {
 
@@ -74,6 +79,7 @@ class Application extends App implements IBootstrap {
 
 	public function boot(IBootContext $context): void {
 		$context->injectFn(Closure::fromCallable([$this, 'registerNotifications']));
+		$context->injectFn(Closure::fromCallable([$this, 'registerCollaborationResources']));
 	}
 
 	public function register(IRegistrationContext $context): void {
@@ -103,9 +109,17 @@ class Application extends App implements IBootstrap {
 		$context->registerEventListener(VoteSetEvent::class, VoteListener::class);
 		$context->registerEventListener(UserDeletedEvent::class, UserDeletedListener::class);
 		$context->registerEventListener(GroupDeletedEvent::class, GroupDeletedListener::class);
+		$context->registerSearchProvider(SearchProvider::class);
 	}
 
 	public function registerNotifications(NotificationManager $notificationManager): void {
 		$notificationManager->registerNotifierService(Notifier::class);
+	}
+	protected function registerCollaborationResources(IProviderManager $resourceManager): void {
+		$resourceManager->registerResourceProvider(ResourceProvider::class);
+
+		\OC::$server->getEventDispatcher()->addListener('\OCP\Collaboration\Resources::loadAdditionalScripts', static function () {
+			Util::addScript(self::APP_ID, 'polls-collections');
+		});
 	}
 }